Create typed ApiResponse type and moved to different namespace
