+++ src/Orchestrator/Controllers/StatusController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
 using Middleware.Common.Repositories;
+using Middleware.Common.Responses;
 using Middleware.Orchestrator.ApiReference;
 
 namespace Middleware.Orchestrator.Controllers;
