+++ src/RedisInterface/Controllers/ActionController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
 using Middleware.Common.Repositories;
 using Middleware.Common.Repositories.Abstract;
+using Middleware.Common.Responses;
 
 namespace Middleware.RedisInterface.Controllers
 {
