+++ src/RedisInterface/Controllers/InstanceController.cs
@@ -3,6 +3,7 @@ using System.Net;
 using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
 using Middleware.Common.Repositories;
+using Middleware.Common.Responses;
 
 namespace Middleware.RedisInterface.Controllers
 {
