+++ src/RedisInterface/Controllers/PolicyController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
 using Middleware.Common.Repositories;
+using Middleware.Common.Responses;
 
 namespace Middleware.RedisInterface.Controllers
 {
