+++ src/RedisInterface/Controllers/RobotController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
 using Middleware.Common.Repositories.Abstract;
+using Middleware.Common.Responses;
 using System.Net;
 
 namespace Middleware.RedisInterface.Controllers
