+++ src/ResourcePlanner/Config/AutoMapperConfig.cs
@@ -1,4 +1,5 @@
 ﻿using Middleware.Common.Models;
+using Middleware.Common.Responses;
 using KeyValuePair = Middleware.Common.Models.KeyValuePair;
 
 namespace Middleware.ResourcePlanner.Config;
