+++ src/ResourcePlanner/Controllers/ReplanResourceController.cs
@@ -4,6 +4,7 @@ using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
 using Middleware.Common.Repositories;
+using Middleware.Common.Responses;
 using Middleware.ResourcePlanner.ApiReference;
 using Middleware.ResourcePlanner.Models;
