+++ src/TaskPlanner/Controllers/PlanController.cs
@@ -2,6 +2,7 @@
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using Middleware.Common.Models;
+using Middleware.Common.Responses;
 using Middleware.TaskPlanner.ApiReference;
 using Middleware.TaskPlanner.Services;
