+++ src/TaskPlanner/Controllers/RePlanController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using Middleware.TaskPlanner.ApiReference;
 using System;
 using YamlDotNet.Core;
+using Middleware.Common.Responses;
 
 namespace Middleware.TaskPlanner.Controllers
 {
